+++ index.d.ts
@@ -1,22 +1,26 @@
 /**
- * react-svg-ionicons v0.2.1 with svg files from Ionicons v4.5.5
+ * react-svg-ionicons v0.2.2 with svg files from Ionicons v4.5.5
  *
  * 317 dual-mode icons, 62 logo icons.
  *
  * @auhtor aMarCruz <amarcruzbox-git@yahoo.com>
- * @date 2019-02-02T11:28:05Z
+ * @date 2019-02-02T13:30:01Z
  * @license MIT
  */
-import * as React from 'react'
+import * as React from 'react';
+
+type Dict<T = any> = { [k: string]: T };
 
 // tslint:disable-next-line:ban-types
-export type Nullable<T> = { [K in keyof T]: (Object extends T[K] ? Nullable<T[K]> : T[K]) | null }
-export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>
+export type Nullable<T> = { [K in keyof T]: (Object extends T[K] ? Nullable<T[K]> : T[K]) | null };
+export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
 
 /**
  * SVG icon renderizer.
  */
-export interface SVGIcon { (opts: object, iconTitle: string, ios?: boolean): JSX.Element }
+export interface SVGIcon {
+  (opts: object, iconTitle: string, ios?: boolean): JSX.Element;
+}
 
 /**
  * IonIcon class properties and attributes.
@@ -26,38 +30,38 @@ export interface IonIconProps<T extends string> extends
   /**
    * The name with which the icon was registered.
    */
-  name: T
+  name: T;
   /**
    * Color for `fill` and `stroke`, in any format accepted by CSS.
    */
-  color?: string
+  color?: string;
   /**
    * Value for `width` and `height`, any format accepted by CSS.
    */
-  size?: string | number
+  size?: string | number;
   /**
    * Icon style, overrides automatic detection of the platform.
    */
-  mode?: 'ios' | 'md'
+  mode?: 'ios' | 'md';
   /**
-   * Overrides the default title (icon name in Title Case).
+   * Overrides the default icon title (icon name in Title Case).
    */
-  title?: string
+  title?: string;
   /**
    * React `ref` to the `<svg>` element.
    */
-  innerRef?: React.Ref<SVGSVGElement>
+  innerRef?: React.Ref<SVGSVGElement>;
 }
 
 /**
  * Bundle with name to icon translations
  */
-export type IconMap = { [K in IconNames]?: SVGIcon }
+export type IconMap = { [K in IconNames]?: SVGIcon };
 
 /**
  * Allows to accept custom names in addIcons.
  */
-export type CustomIconMap = { [k: string]: SVGIcon | null | undefined }
+export type CustomIconMap = Dict<SVGIcon | null | undefined>;
 
 /**
  * Possible defaults
@@ -68,7 +72,7 @@ export interface IonIconDefs extends Omit<IonIconProps<''>, 'name' | 'title' | '
 /**
  * Size aliases
  */
-export type IonIconSizes = { [k: string]: string | number | undefined }
+export type IonIconSizes = Dict<string | number | undefined>;
 
 /**
  * Main component to render a SVG ionicon.
@@ -83,7 +87,7 @@ export declare class IonIcon extends React.PureComponent<IonIconProps<IconNames>
  *
  * @param iconMap Object with name-icon translations.
  */
-export function addIcons(iconMap: Nullable<IconMap> & CustomIconMap): void
+export function addIcons(iconMap: Nullable<IconMap> & CustomIconMap): void;
 
 /**
  * Merge the given values with the current defaults.
@@ -92,7 +96,7 @@ export function addIcons(iconMap: Nullable<IconMap> & CustomIconMap): void
  *
  * @param defaults Properties to merge.
  */
-export function setDefaults(defaults: Nullable<IonIconDefs>): void
+export function setDefaults(defaults: Nullable<IonIconDefs>): void;
 
 /**
  * Reset the table of named sizes.
@@ -102,7 +106,22 @@ export function setDefaults(defaults: Nullable<IonIconDefs>): void
  *
  * @param sizes Object with a sizes map.
  */
-export function setSizes(sizes: Nullable<IonIconSizes>): void
+export function setSizes(sizes: Nullable<IonIconSizes>): void;
+
+/**
+ * Add icon titles to the icon-name -> title transtations.
+ *
+ * @param sizes Object with a {icon-name: title} props.
+ */
+export function setTitles(iconTitles: Dict<string | null>): void;
+
+/**
+ * Class or space separated classes to add to the `className` property
+ * of all the icons, in addition to the default or specific one.
+ *
+ * @param class Class or space separated list of classes.
+ */
+export function setBaseClass(classes: string): void;
 
 /**
  * Icon names
@@ -487,3 +506,4 @@ export type IconNames =
   | 'wifi'
   | 'wine'
   | 'woman'
+;